Refactor and reorganize code structure; move experimental SSL inline patching to a separate directory, update README documentation, and enhance logging for better traceability.
