+++ .gitignore
@@ -27,3 +27,4 @@ copydll.bat
 .github
 .vscode
 .idea
+GEMINI.md
