+++ experimental/README.md
@@ -1,0 +1,7 @@
+# Experimental Code
+
+This directory contains experimental, unstable, or incomplete features that are **not** part of the production build.
+
+Code in this directory is kept for research and development purposes, or as a potential fallback if primary methods fail in the future.
+
+**DO NOT USE THIS CODE IN PRODUCTION.**
